Update FormulaCreator.java
stage0-module3-task4 second attempt

+++ src/main/java/lang/print/gaps/task4/FormulaCreator.java
@@ -2,8 +2,8 @@ package lang.print.gaps.task4;
 
 public class FormulaCreator {
     public static void main(String[] args) {
-    	short a = 5;
-    	short b = 2;
+    	double a = 5;
+    	double b = 2;
     	double c;
     	
     	c = (9*a*a-5*b+2+a-7)*((a+b-4*a*b)/2);
